Incorrect reference to FirebaseAuth example
I believe this reference to the FirebaseAuth example has been forgotten to be removed from the README following a change that now uses StyledFirebaseAuth in the example only. This is to correct the doc (simply remove the reference).

Would be nice to actually see an example implenentation of FirebaseAuth, though.